@@ -6,16 +6,19 @@
     "dev": "next dev",
     "build": "next build",
     "pages:build": "next build && npx @cloudflare/next-on-pages",
+    "pages:dev": "npx @cloudflare/next-on-pages --watch & npx wrangler pages dev .vercel/output/static --compatibility-flag=nodejs_compat",
     "start": "next start"
   },
   "dependencies": {
-    "next": "14.2.35",
+    "next": "15.5.2",
     "react": "18.3.1",
     "react-dom": "18.3.1"
   },
   "devDependencies": {
     "typescript": "5.9.3",
     "@types/node": "20.19.27",
-    "@types/react": "18.3.27"
+    "@types/react": "18.3.27",
+    "wrangler": "^3",
+    "@cloudflare/next-on-pages": "^1"
   }
 }
