@@ -1,4 +1,6 @@
+/** @type {import('next').NextConfig} */
 const nextConfig = {
-  typescript: { ignoreBuildErrors: true },
+  reactStrictMode: true
 };
-module.exports = nextConfig;
+
+export default nextConfig;
